+++ Software/workspace/SimCanTest/src/Subsystems/Shooter.h
@@ -9,6 +9,8 @@
 #define SRC_SUBSYSTEMS_SHOOTER_H_
 
 #include "WPILib.h"
+#include "CANTalon.h"
+
 
 class Shooter: public Subsystem , public PIDSource, public PIDOutput {
 	CANTalon angleMotor;
@@ -60,7 +62,7 @@ public:
 
 	void Execute();
 	void Reset();
-	void PIDWrite(float output);
+	void PIDWrite(double output);
 
 };
